Extract the entropy source utilities to a separate file
