+++ src/index.ts
@@ -1,6 +1,7 @@
-import { x64hash128 } from './utils/hashing'
 import { load, Agent, LoadOptions, GetOptions, GetResult, hashComponents, componentsToDebugString } from './agent'
-import { Component, UnknownComponents, BuiltinComponents } from './sources'
+import { BuiltinComponents } from './sources'
+import { Component, UnknownComponents } from './utils/entropy_source'
+import { x64hash128 } from './utils/hashing'
 
 // Exports that are under Semantic versioning
 export {
@@ -22,6 +23,9 @@ export default { load, hashComponents, componentsToDebugString }
 // The exports below are for private usage. They may change unexpectedly. Use them at your own risk.
 /** Not documented, out of Semantic Versioning, usage is at your own risk */
 export const murmurX64Hash128 = x64hash128
+export { prepareForSources } from './agent'
+export { sources } from './sources'
+export { getScreenFrame } from './sources/screen_frame'
 export {
   getFullscreenElement,
   isTrident,
@@ -31,5 +35,4 @@ export {
   isGecko,
   isDesktopSafari,
 } from './utils/browser'
-export { getScreenFrame } from './sources/screen_frame'
-export { loadSources, SourcesToComponents } from './sources'
+export { loadSources, SourcesToComponents } from './utils/entropy_source'
